+++ src/Control/Eff/Concurrent/Protocol/StatefulServer.hs
@@ -82,19 +82,19 @@ class (Typeable (Protocol a)) => Server (a :: Type) q where
   -- | Return an initial 'Model' and 'Settings'
   setup ::
        StartArgument a q
-    -> Eff (Processes q) (Model a, Settings a)
+    -> Eff q (Model a, Settings a)
 
   default setup ::
        (Default (Model a), Default (Settings a))
     => StartArgument a q
-    -> Eff (Processes q) (Model a, Settings a)
+    -> Eff q (Model a, Settings a)
   setup _ = pure (def, def)
 
   -- | Update the 'Model' based on the 'Event'.
   update ::
        StartArgument a q
     -> Effectful.Event (Protocol a)
-    -> Eff (ModelState a ': SettingsReader a ': Processes q) ()
+    -> Eff (ModelState a ': SettingsReader a ': q) ()
 
 -- | This type is used to build stateful 'EffectfulServer' instances.
 --
@@ -103,10 +103,10 @@ class (Typeable (Protocol a)) => Server (a :: Type) q where
 -- @since 0.24.0
 data Stateful a deriving Typeable
 
-instance Server a q => Effectful.Server (Stateful a) (Processes q) where
-  data Init (Stateful a) (Processes q) = Init (StartArgument a q)
+instance Server a q => Effectful.Server (Stateful a) q where
+  data Init (Stateful a) q = Init (StartArgument a q)
   type ServerPdu (Stateful a) = Protocol a
-  type ServerEffects (Stateful a) (Processes q) = ModelState a ': SettingsReader a ': Processes q
+  type ServerEffects (Stateful a) q = ModelState a ': SettingsReader a ': q
 
   runEffects (Init sa) m = do
     (st, env) <- setup sa
@@ -124,9 +124,9 @@ start
     , Typeable a
     , LogsTo h (Processes q)
     , Effectful.Server (Stateful a) (Processes q)
-    , Server a q
+    , Server a (Processes q)
     )
-  => StartArgument a q -> Eff (Processes q) (Endpoint (Protocol a))
+  => StartArgument a (Processes q) -> Eff (Processes q) (Endpoint (Protocol a))
 start = Effectful.start . Init
 
 -- | Execute the server loop.
@@ -138,9 +138,9 @@ startLink
     , Typeable a
     , LogsTo h (Processes q)
     , Effectful.Server (Stateful a) (Processes q)
-    , Server a q
+    , Server a (Processes q)
     )
-  => StartArgument a q -> Eff (Processes q) (Endpoint (Protocol a))
+  => StartArgument a (Processes q) -> Eff (Processes q) (Endpoint (Protocol a))
 startLink = Effectful.startLink . Init